requirements.txt
PyYAML~=6.0
beautifulsoup4~=4.12.1
requests~=2.28.2
urllib3~=1.26.15
